requirements.txt
chromadb>=0.5.0
google-genai>=1.0.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
prompt-toolkit>=3.0.0
rich>=13.0.0
kokoro>=0.9.4
sounddevice>=0.5.0
soundfile>=0.13.0
httpx>=0.27.0
miniaudio>=1.59
fastapi>=0.115.0
uvicorn>=0.34.0
sse-starlette>=2.0.0
